fix: rename image files
